+++ Arduino/libraries/commandData/commandData.h
@@ -6,23 +6,23 @@
 
 typedef struct
 {
-    float yaw_ccw;
-    float throttle_up;
-    float roll_left;
-    float pitch_forward;
-    float mode_button;
-    float servo_button;
-    float autoland;
+    double yaw_ccw;
+    double throttle_up;
+    double roll_left;
+    double pitch_forward;
+    double mode_button;
+    double servo_button;
+    double autoland;
     #ifdef GAIN_TUNING
-        float roll_kp;
-        float roll_ki;
-        float roll_kd;
-        float pitch_kp;
-        float pitch_ki;
-        float pitch_kd;
-        float yaw_kp;
-        float yaw_ki;
-        float yaw_kd;
+        double roll_kp;
+        double roll_ki;
+        double roll_kd;
+        double pitch_kp;
+        double pitch_ki;
+        double pitch_kd;
+        double yaw_kp;
+        double yaw_ki;
+        double yaw_kd;
     #endif
 } commanddata_t;
